+++ src/SAMD_Slow_PWM.h
@@ -12,13 +12,14 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.0
+  Version: 1.2.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      01/10/2021 Initial coding for SAMD21/SAMD51 boards
   1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
   1.2.0   K Hoang      31/01/2022 Fix multiple-definitions linker error. Improve accuracy. Change DutyCycle update
+  1.2.1   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period
 *****************************************************************************************************************************/
 
 #pragma once