+++ src/SAMD_Slow_PWM_ISR_Impl.h
@@ -12,13 +12,14 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.0
+  Version: 1.2.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      01/10/2021 Initial coding for SAMD21/SAMD51 boards
   1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
   1.2.0   K Hoang      31/01/2022 Fix multiple-definitions linker error. Improve accuracy. Change DutyCycle update
+  1.2.1   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period
 *****************************************************************************************************************************/
 
 #pragma once
@@ -116,7 +117,7 @@ void SAMD_Slow_PWM_ISR::run()
           PWM[channelNum].period    = PWM[channelNum].newPeriod;
           PWM[channelNum].newPeriod = 0;
           
-          PWM[channelNum].onTime  = ( PWM[channelNum].period * PWM[channelNum].newDutyCycle ) / 100;
+          PWM[channelNum].onTime  = PWM[channelNum].newOnTime;
         }
 #endif
       }    
@@ -152,13 +153,12 @@ int SAMD_Slow_PWM_ISR::findFirstFreeSlot()
 
 ///////////////////////////////////////////////////
 
-int SAMD_Slow_PWM_ISR::setupPWMChannel(const uint32_t& pin, const double& period, const double& dutycycle, void* cbStartFunc, void* cbStopFunc)
+int SAMD_Slow_PWM_ISR::setupPWMChannel(const uint32_t& pin, const uint32_t& period, const float& dutycycle, void* cbStartFunc, void* cbStopFunc)
 {
   int channelNum;
   
   // Invalid input, such as period = 0, etc
-  //if ( (period == 0) || (dutycycle > 100) )
-  if ( (period <= 0.0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
+  if ( (period == 0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
   {
     PWM_LOGERROR("Error: Invalid period or dutycycle");
     return -1;
@@ -193,10 +193,10 @@ int SAMD_Slow_PWM_ISR::setupPWMChannel(const uint32_t& pin, const double& period
   PWM[channelNum].callbackStart = cbStartFunc;
   PWM[channelNum].callbackStop  = cbStopFunc;
   
-  PWM_LOGDEBUG0("Channel : ");      PWM_LOGDEBUG0(channelNum); 
-  PWM_LOGDEBUG0("\t    Period : "); PWM_LOGDEBUG0(PWM[channelNum].period);
-  PWM_LOGDEBUG0("\t\tOnTime : ");   PWM_LOGDEBUG0(PWM[channelNum].onTime); 
-  PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
+  PWM_LOGINFO0("Channel : ");      PWM_LOGINFO0(channelNum); 
+  PWM_LOGINFO0("\t    Period : "); PWM_LOGINFO0(PWM[channelNum].period);
+  PWM_LOGINFO0("\t\tOnTime : ");   PWM_LOGINFO0(PWM[channelNum].onTime); 
+  PWM_LOGINFO0("\tStart_Time : "); PWM_LOGINFOLN0(PWM[channelNum].prevTime);
  
   numChannels++;
   
@@ -207,10 +207,10 @@ int SAMD_Slow_PWM_ISR::setupPWMChannel(const uint32_t& pin, const double& period
 
 ///////////////////////////////////////////////////
 
-bool SAMD_Slow_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const double& period, const double& dutycycle)
+bool SAMD_Slow_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const uint32_t& period, const float& dutycycle)
 {
   // Invalid input, such as period = 0, etc
-  if ( (period <= 0.0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
+  if ( (period == 0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
   {
     PWM_LOGERROR("Error: Invalid period or dutycycle");
     return false;
@@ -232,11 +232,12 @@ bool SAMD_Slow_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const
 
   PWM[channelNum].newPeriod     = period;
   PWM[channelNum].newDutyCycle  = dutycycle;
+  PWM[channelNum].newOnTime     = ( period * dutycycle ) / 100;
   
-  PWM_LOGDEBUG0("Channel : ");      PWM_LOGDEBUG0(channelNum); 
-  PWM_LOGDEBUG0("\tNew Period : "); PWM_LOGDEBUG0(PWM[channelNum].newPeriod);
-  PWM_LOGDEBUG0("\t\tOnTime : ");   PWM_LOGDEBUG0(( period * dutycycle ) / 100); 
-  PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
+  PWM_LOGINFO0("Channel : ");      PWM_LOGINFO0(channelNum); 
+  PWM_LOGINFO0("\tNew Period : "); PWM_LOGINFO0(PWM[channelNum].newPeriod);
+  PWM_LOGINFO0("\t\tOnTime : ");   PWM_LOGINFO0(PWM[channelNum].newOnTime); 
+  PWM_LOGINFO0("\tStart_Time : "); PWM_LOGINFOLN0(PWM[channelNum].prevTime);
   
 #else
 
@@ -249,10 +250,10 @@ bool SAMD_Slow_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const
   
   PWM[channelNum].prevTime      = timeNow();
    
-  PWM_LOGDEBUG0("Channel : ");      PWM_LOGDEBUG0(channelNum); 
-  PWM_LOGDEBUG0("\t    Period : "); PWM_LOGDEBUG0(PWM[channelNum].period);
-  PWM_LOGDEBUG0("\t\tOnTime : ");   PWM_LOGDEBUG0(PWM[channelNum].onTime); 
-  PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
+  PWM_LOGINFO0("Channel : ");      PWM_LOGINFO0(channelNum); 
+  PWM_LOGINFO0("\t    Period : "); PWM_LOGINFO0(PWM[channelNum].period);
+  PWM_LOGINFO0("\t\tOnTime : ");   PWM_LOGINFO0(PWM[channelNum].onTime); 
+  PWM_LOGINFO0("\tStart_Time : "); PWM_LOGINFOLN0(PWM[channelNum].prevTime);
   
 #endif
   
@@ -263,13 +264,8 @@ bool SAMD_Slow_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const
 
 void SAMD_Slow_PWM_ISR::deleteChannel(const uint8_t& channelNum) 
 {
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
-
   // nothing to delete if no timers are in use
-  if (numChannels == 0) 
+  if ( (channelNum >= MAX_NUMBER_CHANNELS) || (numChannels == 0) )
   {
     return;
   }